Sea-ice:
- as if we didn't have enough free parameters, I have added some
functionality to allow for variable amounts of lateral (side) melt for
sea ice. The new database parameter 'silmfac' controls this
strength. The default value is 2.5d-8 (taken from published values),
but could vary from 0 to 2 or 3 times as much. Note that this does not
affect the fixed-SST runs, but can be used to tune a q-flux run if
necessary...

- New seaice conservation points have been added so that we can better
distinguish between Lat Melt/Advection/thermodynamics/ocean formation
etc.

- ICEDYN code has been slightly altered to use the standard TRIDIAG
solver (rather than it's own), and some arrays rearranged for better
efficiency.