New albedo algorithm for snow covered soil/vegetation is introduced. It uses snow fraction which is computed according to topographic properties of the surface (standard deviation of the elevation).
The snow fraction is computed in GHY_DRV.f : snow_cover()

New database parameter is introduced : snow_cover_coef . It has
default value of .15 and controles the scale at which topography is
used for computation of the snow cover. If set to 0 then everything is
flat and topography is neglected.