To make things more transparent and to improve portability of modelE the setup script was rewritten in PERL. Perl has a unique standard, so the script should run on any machine on which perl is installed. I tried to reproduce exactly the same functionality as in the old "setup", but could miss something. Please let me know if you think that something in setup is not working properly.
New option is added to modelErc: UMASK. UMASK defines the value
of 'umask' you want to use for file permissions inside the run
directory. New option allows people to choose the level of security
depending on the machine they are working on. The default is 002 which
seems reasonable for GISS.