CLOUDS:
Fix of a very rare error when downdrafts sometimes detrain because
they are too large. This does not have a big impact on climate, but it
was important for tracers.

For the tracers, I have added a few variables that needed to be passed
for calculating tracer amounts in condensate and evaporation. Currently these
are only used for TRACERS_SPECIAL_O18.

Some new physics has been added for the isotopes (kinetic
fractionation when condensing to ice, evporating below clouds and only
partial equilibration for moist convective rain (since it has bigger
drops).

TRACERS_DRV:

Adjustments to the GET_COND, GET_EVAP factors.

NOTE: previously we had assumed that condensation was to ice or water
as a function of local temperature (i.e. ice below freezing). This is
not actually what is going on. For the isotopes, this is now decided
by LHX, possibly this should be done as well for other tracers?

TRACERS_O18:

New kinetic fractionation calculations