The problem with the land ice albedo that I committed yesterday was because the 2-band variables that I assumed were synonymous with the 6 band variables were actually being used sneakily to pass values to the Ant/Grnld albedo specification.
This lead me to find an error in the original code where the snowage
for land snow was being used for land ice (and ocean ice), but only in
the 6-band case. Correcting this has no effect since by default we run
with equal snowages (but there are sets of options that would have
caused a problem).

So, I have made some further simplifications/clarifications to the
code to fix both problems.

As before, the answers should be identical as before (definitely true
for KVEGA6=0 and 3, will not be true for KVEGA6=2).

Note that there is still much single precision in the radiation
code. At some point that will be remedied....