Tracer corrections (TRACERS_WATER only)
OCEAN.f: sea ice tracers were being given wrong default

TRACERS_DRV.f: round-off error fixed in initialisation of lakes

SEAICE.f:
- a number of places where we wern't careful enough with tracer fluxes
  have been fixed.
- very rare problem of DEW and TRDEW not being the same sign is now dealt
  with
- new check to find negative tracer values (it was needed!)

The only change that effects the main model is the change in
ADDICE. At some points we round up the RSI value if it is close (<
1d-4) to it's max. We were incorrectly redistributing the mass, so
that tracers (as well as heat and salt) were not being properly
conserved. Now they are, but I doubt this will make a big difference
to the results.