New options for calculating orbital parameters:
I have included the basic Berger (1978) routine for calculating
orbital parameters into RAD_DRV.f. If the parameter calc_orb_yr = 1,
the parameter paleo_orb_yr (in years before present = 1950) will be
used to calculate the eccentricity, obliquity, and precession for that
year. Note that due to minor truncation effects, the values given by
this routine, will not be identical to those used, for instance, in
the PMIP experiments.  If new fixed values are desired, the defaults
could be changed in CONST.f. Conceivably, we could in future allow the
orbital calculation to be adjusted annually...

The new call to ORBPAR is in init_rad, and the subroutine and also
ORBIT have been put into RAD_DRV.f.

PS. the difference from our default values in the insolation if you
take paleo_orb_year as -50, (ie. 2000 AD), is 0.01 W/m^2 in the global
mean (for the first hour).