New documentation file: modelE.html
This will become the ModelE reference manual. Currently the only bits
that are filled in are the descriptions and explanations of the
directory strucure and notes on the contents. If you look at the file
you will notice that there is a strucure set up that we can fill in
bit by bit explaining what goes on in the code.

It will be greatly appreciated if people who have responsibility for
particular bits of the code, were able (at minimum) to give maybe a
recent reference and any modifications made since then.

I would hope to have this document morph into a full publication at
some point and obviously all contributors will be acknowledged.