New data base parameter: dt_UVfilter The effect of the UV-filter is multiplied by dt/dt_UVfilter , hence becomes independent of the dynamics time step dt.
To reproduce the previous model, set dt_UVfilter=DT .

However, it turned out that the UV-filter in the current model is not
needed if we apply the sea level pressure filter every hour rather than
every 2 hours. Therefore the defaults were changed to that case:
NFILTR=1   and   dt_UVfilter=0.

U00ice had to be changed to .53 to get radiative balance.