Ice model/Fluxes change:
1) New atmospheric field 'APRESS' now available for ocean models (in
FLUXES module) which is the integrated gauge pressure at the surface
of the ocean (and under sea ice).

Ocean model changes:

2) To prevent (usually v. small) oscillatory pattern in sea surface
height adversely affecting ice dynamics, the OGEOZA is now the average
of the last two time steps. New array OGEOZ_SV is now in restart file.

3) Correction to stresses applied to ocean

4) Updating of surface ocean pressure within ocean code takes account
of ice fluxes 'in transit' so to be speak.