Modification to fluxes from lateral melt of sea ice.
- lateral melt now affects lakes that are partially ice covered

- Because of a quirk in the qflux model, the combination of river flow
fluxes and lateral melt fluxes has had to be changed. Therefore the
fluxes FLOWO and EFLOWO revert to being pure river flow fluxes, and
the new fluxes MELTI, EMELTI and SMELTI are now the fluxes from the
lateral melt of sea ice. These new arrays (like E/FLOWO) are in kg and
J and should be applied to the whole grid box.

- The vertical fluxes of energy into the ocean needed for the qfluxes
have been corrected (so as not to include lateral sea ice melt).

- The qflux model itself now takes special care for the lateral melt in
a way that should now conserve energy.

- some minor changes to variable names to be more intuitive