Minor corrections to qflux/ice model
- meltsi is moved to before clouds and radiation for better
consistency in the diagnostics. Only minor changes to results as far
as I can tell.

- correction of minor in the qflux calculation when precip is
added to the ocean. This crept in a while ago when we pulled PRECIP
apart and changed the order of calculation from ocean then ice, to ice
then ocean.

- move the addition of lateral ice melt to precip calculation rather
than ground for slightly improved consistency.

None of these errors seem to make much difference to either the
control run or the basic qflux model.