Diagnostic changes:
These changes all arise from checks to the needed correspondence
between the budget tables and the conservation diagnostics. Some AJ
diags were corrected (notably the ice melt diagnostics). In order for
easier comparison, I have added two more conservation diagnostics so
that lake and ocean ice mass and energy can be seen separately and
more easily compared.

Note as well that the Lake Mass and Energy conservation diagnostics were
significantly wrong and are now corrected.

There will be a new file conserv.txt in the ../doc directory that
outlines the various correspondences so future changes can be measured
against these conservation principles. There may be more, so feel free
to add ones you think are useful.

Note that it appears the ground hydrology is not properly conserving,
but we are working on fixing that.