conserv.txt:
New document outlining the conservation principles that should be true
in the model. Note that if the numbers are extracted from the PRT
file, there can be significant round off error. Also, unless NRAD=1,
the energy numbers will not work for the first hour print out.

Does anyone want to convert this into a diagnostic routine so that we
can automatically keep a check on mass, salt, and energy conservation?