Big fix for 18 layer (and strat models).
The phase of cloud formation is in part determined by the phase of
immediately overlying cloud water (through LHXUP). In models that go
up to the stratosphere, the warm termperatures at high altitude caused
LHXUP to be set for water clouds. This was then propagated down
regardless of whether any clouds were actually formed. Net result is
that no ice clouds are formed if the stratosphere is warm. This error
affected all model II' runs that include the stratosphere.

The fix is to only allow LHXUP to influence LHX if there is actually
cloud in the above layer (WMX(L+1)>0). Possibly this test may need to
be refined in future (Yao?).

In the 12 layer model, this has only a very minor effect, but for the
18 layer model, the effects are very significant. All tunings will
need re-doing. The difference between 12 and 18 layer models are now
much reduced when starting from identical initial conditions.

I made an auxilliary change to the setting of the random numbers in
RAD_DRV.f so that changes to LSCOND do not lead to changes in the
random numbers used for moist convective clouds.