Further modifications to allow coarse and fine resolution models
Parallel fix to FFT144 which somehow got missed previously

STRATDYN has a minor difference in the EK array

GHY_DRV used to mysteriously read in vdata(11) (which is not an IM,JM
array), and which does not work with higher resolution. Does anyone
know what this was for?

New module FFT36 for the coarse resolution runs

New rundeck, 12 layer, 2x2.5 model. Compiles and runs, but obviously
no-one has tested the tuning (SDRAG, U00 etc.).

If anyone wants to test this, please go ahead!