New variable FWSIM(IM,JM) (fresh water sea ice mass) is the integrated amount of freshwater in the ocean and lake ice.
In order to conserve freshwater, the mixed layer ocean (which is
assumed to be fresh) can only take account of the freshwater amount in
the sea ice when calculating the mass balance. Hence, the qflux code
has been adjusted to use FWSIM instead of ACE1I+MSI+SNOW. The mass
fluxes into the qflux ocean are similarly adjusted to only represent
freshwater (i.e. no salt).

This conserves better, but is slightly more complicated. For the qflux
calculations some adjustments need to be made, and so the OA array is
also slightly redefined. Some minor re-arrangements of the
vertflux/ocnIC/otspec code was necessary.

Please examine the output of these calculations carefully and let me
know if there are any problems.