Qflux model can now have ice advection:
As further option, the qflux model can be run with sea ice
dynamics. By choosing to put

ICEDYN_DUM                ! dummay ice dynamics

or

ICEDYN_DRV ICEDYN         ! ice dynamics

in the rundeck for the spin-up run (KOCEAN=0), the model will
calculate the heat convergence from ice dynamics and save it as OA(13)
for the qflux calculations.

The same rundeck configuration can be used for the qflux model, for
which the ice advection will occur as if for a fully coupled model.

Note that you now MUST include one of the two options. ICEDYN_DUM
contains the code for the windstress-dependent oi_ustar, which is no
longer a useful variable if full ice dynamics are used.

I have done one sneaky thing in incorporating ADVSI_DIAG which is
needed to do proper accounting of the implicit mass and heat fluxes in
the qflux ocean. This is called from ADVSI, and is a dummy routine in
the same file. Only if the real routine (in OCEAN.f) is linked prior
to linking the ice dynamics will this do anything. i.e. for full
oceans, it is a dummy routine, but for the qflux ocean OCEAN MUST come
before ICEDYN_DRV. Sorry about that!

ICEDYN_DRV: Minor modification to adjust the sea surface height
forcing to depend on how much the free surface is elevated by ice
displacement.

OCEAN: Minor corrections to extreme case when ice is deeper than mixed
layer.