Changed loops over bare and vegetated soil so that if fraction of bare or vegated soil is == 0, then the computations are not performed for the corresponding emtpy part of the cell. This is done because sometimes computations inside those "empty" fractions caused instabilities. They were also wasting cpu cycles.
The results may slightly change because previously the computations
of the GHY time step were taking into account conditions inside those
"empty" fractions, which is not done now.
If you run into problems with this version of GHY.f , try to comment
out the lines
      if ( fb == 0.d0 ) then  ! bare fraction is missing
        i_bare = 2
        process_bare = .false.
      endif
      if ( fv == 0.d0 ) then  ! bare fraction is missing
        i_vege = 1
        process_vege = .false.
      endif
This should effectively return it to the older algorithm.