Couple fixes to dynaic vegetation code.
"Cin" was not passed out of subroutine "cond" causing zero division.

"s" was not set in "qsimp" causing floating point exception. I set
it to 1 since I have no idea what it is. This data is not yet passed
to other parts of the code, so hopefully it is ok for now.