Fixed an error in the sea surface tilt calculation for ice dynamics which caused ice dynamics to be unstable for coupled models. It was relatively benign for the qflux model.
However, this fix might now have adverse consequences for the qflux
model running with ICEDYN. Thus I have added an option (osurf_tilt) to
change the tilt term calculation.

If osurf_tilt = 1, then the full calculation is done based on the free
surface height (as it should be for fully coupled models, where it is
the default). However, for the qflux models, this is set to 0 and so
the calculation is done from geostrophy. Since qflux ocean currents
are zero, this implies that the sea surface tilt is zero.

Other coupled models need to upgrade to this fix, but do not need to
adjust their code.