Modularisation of Glacial (Iceberg) flux into ocean:
Documentation of New fluxes:

GMELT(IM,JM) (kg), EGMELT(IM,JM) (J) are the mass and energy
associated with the addition of icebergs into the ocean around
Antarctica (Weddell and Ross Seas) and Greenland. The amounts are
the annual means, based on IPCC SAR estimates of these fluxes.

These fluxes can be treated exactly like river runoff (FLOWO, EFLOWO)
if desired, or alternatively, can be distributed over depth (say 400m)
corresponding to the mean draft of the icebergs themselves. This can
add mass under the mixed layer.

The water is added with negative energy as if it was ice (i.e. the
energy needed to melt icebergs comes from the ocean). This may lead to
below freezing conditions in the affected boxes.

Diagnostics for these fluxes are put in with the river runoff diags.

Please note that the placing of the iceberg flux is rather arbritrary,
and thus the coding is quite sensitive to the resolution. If radically
different landmasks are used, the code setting GMELT (in init_LI:
LANDICE_DRV.f) should be adjusted accordingly. Currently supported
resolutions (4x5, 2x2.5) will work fine. For 8x10 resolutions, this
flux is turned off.