Slight modification to the glacial (iceberg) melt code. The default is now that this is turned OFF. Thus previous version (2.3.1) fixed SST and qflux models are still valid.
To turn it on (which should have no effect on qflux models, but big
impacts in fully coupled ocean models), set 'glmelt_on=1' in the
rundeck. THIS NEEDS TO BE DONE FOR ALL COUPLED MODELS THAT WANT THIS
FUNCTIONALITY.

The coupled model rundeck for E001o.R is suitably updated.