Adjustment of some diagnostics to provide a cleaner diagnosis for energy conservation in sea ice.
Now implicit heat and mass flux diags only apply to fluxes at the base
of the ocean mixed layer. For fixed SST runs, they apply at the base
of the ice.

Note that there was a possibly serious error in the setting of
AIJ(J_SMFX) in the qflux model when J_SMFX was not defined. This only
occurred if the mixed layer depth exceeded the maximum
though. Probably this just affected another diagnostic, but I can't be
sure.