Correction of some minor initialisation errors to do with sea ice salinity.
Also, the GIC file: GIC.rsfB357M12.1DEC1956.1 does not appear to have
been created properly and all sea ice salinities are zero. This is not
wrong necessarily, but this will lead to unecessary drift in the qflux
or coupled models. I have therefore replaced all mentions of this file
with the extended version, which does not appear to have this problem.

A correction is also made to the implicit heat flux diagnostic on the
budget page. Please check that other models that are run (particularly
q-flux) do indeed conserve energy of sea ice. (see doc/conserv.txt for
details).