CMPE002 is just more convenient implementation of CMPE001. It uses modelE reading utilities and gets data through USE instructions. So one doesn't need to modify it when restart file changes. It also works with restart files that contain tracers. I wrote it for myself some time ago, but since some are still using CMPE001, I thought that I better commit it. I guess I had to do it earlier (sorry Gavin).
Only a small number of arrays are checked at the moment, but it is very
easy to add the new ones, just add the line:
   check("name",array)
and include a corresponding USE statement.
I am using a macro definition, so the code had to be in free-form. I hope
that nobody minds since this is just an auxiliary program.
It was tested on SGI and Linux. If one wants to run it on DEC one has to
set corresponding options for the free-form code.