Generic coding for Reha's sub-gridscale wind distribution.
This needs information about downdraft mass fluxes in level 1 from
clouds (hence new array DDM1 - in both clouds versions), and many
surface flux related quantities available within the PBL
calculation. The calculation is different for each itype (land, ocean,
ice etc.).

Currently the functional code is commented out, but an example is laid
out for how to use this within the PBL. Please note, this code
requires inputs that are only available within the PBL code, it will
be easiest and most consistent if people use this here, rather than
after the whole surface calculation.

The calculation of 'sig' the distribution parameter is independent of
tracer (but not t, q or ws) and simply needs to be uncommented. For
each tracer that has a particular wind dependence and threshold
velocity, and integration over wind speed can be performed by calling
'integrate_sgswind'. I have built in simple power law dependent
functions (w, w^2, w^3 etc.), but anything more complicated can be
easily added. The intgreated value 'wint' is the one that should be used in
any tracer boundary condition, and is passed back up to the SURFACE
and GHY_DRV code. I think this needs to multiply ws to get the
effective flux (Reha?).

Conceivably this could be used for temperature and moisture fluxes
too, but then sig would need to be part of the iteration. Not sure
what functional dependence would be used though.

Please let me know if there are any problems with this, or whether
something is unclear in how to use it.