Bug fix for dynamic vegetation call (i.e. cond_scheme=2). "qv" was used in "veg" before it was initialized (in "evap_limits"). This should fix the OpenMP problem we had with this code.
Also added "parameter" attribute to some declarations in vegetation code.
Actually one should always use "parameter" attribute for constants.
The developers of F90 did their best to confuse the programmers.
The fortran declaration
   real a = 1.
has completely different meaning then C declaration
   float a = 1.;
While in C it will allocate "a" on stack and initialize it on each call
to the subroutine, in fortran it will allocate static memory for it
and initialize it only once at the start of the program, i.e. it
is equivalent to
   static float a = 1.;
So the use of "parameter" for constants really helps to avoid the confusion.