This is an attempt to separate all vegetation code to a separate module. Three new files were introduced:   VEGETATION.f VEG_COM.f VEG_DRV.f in analogy with ground hydrology code.   VEGETATION.f contains physics (currently canopy conductance)   VEG_COM.f    contains data specific to vegatation   VEG_DRV.f    initializes the module at the start of the program                and on each time step
The results are identical to those before the separation (at least for
E001.R). For obvious reasons I couldn't test it for tracers. The last
time I was testing it with tracers I had a problem with diagnostic
output. Reto suggested KDIAG(3)=0 as a temporary fix. Please let me
know if you find any problems with this new module.