A re-formulation of the new CLOUDS2 scheme so that it actually works as advertised.
The fraction of the box involved in moist convection and subsidence is
now calulated prior to the determination of plume properties, and all
fluxes are now scaled by this fraction. This implies that the whole
calculation occurs almost exactly as before (i.e. CLOUDS.f) but the
relative fraction of mass in the plume is proportionately higher.

I have also temporarily fixed a small energy conservation error.

Please note that although some changes have been made to the tracer
code in light of the CLOUDS2 algorithm, this is far from complete.

Some new constants defined, and better diagnostics of crashes. I have
reinstated a hard model crash for negative Q since this should never
happen. If it does, it must be a code error which will require fixing
prior to restarting the model.

This code has been tested sucessfully for a month, but I leave any
re-tuning, validation or comparison with previous coding to others.