In an attempt to clean up the code I was trying to find an option in F90 compiler which would give a warning if some variables are declared but not used. It seems that such option doesn't exist :-(.
So, I modified gcmdoc.pl to do some primitive check for unused variables.
Now if you create the documentation with "gmake htmldoc RUN=...", all
the variables which are suspected to be declared but not used will
appear in bright red. It is rather primitive check, so sometimes it will
highlight the variable which is actually used, like variables that are
declared in one module, but "used" from the other one, or when the
variable is used only in declarations (like dimensions of the arays).
Still, it provides a good tool for cleaing the code. I suggest that
everybody looks at the dynamic documentation for the programs they
are working with to see if some variables should be removed.
For the rundeck E001.R the dynamic documentation is updated every
day and is available at :

   http://simplex.giss.nasa.gov/modelE/decks/E001ia_doc/index.html

Please, be aware that this tool gives just a hint. You should carefully
check it yourself if any variable can be removed.

Processing of "entry" was removed from the script because it was conflicting
with the way the data for variables is collected. One shouldn't use
"entry" anyway since it is a bad coding practice,