minor adjustment to allow TRACER_SPECIAL_Lerner to optionally run with Water.
Template rundeck E001tr.R is changed to include the stratosphere, and
a wider variety of plug-and-play options. (4th order momentum, ATRUB
and no ice dynamics). This is being used as a complementary rundeck in
the GSFC test suite, and therefore needs to be as orthogonal as
possible. If you use this rundeck in other circumstances, be aware
that it is not the same configuration as before!