This is a preliminary guide to understanding what is happening to the model as we progress towards ESMF compliance. Principally, it is related to the domain decomposition which is being implemented by collaborators at GSFC.
As this process continues, GISS developers are going to have to pay
more attention to how they declare, allocate and initialise arrays
that extend over latitude and longitude. Some examples have already
been committed in the CLOUDS_COM/CLOUDS_DRV and ICEDYN_DRV/ICEDYN
files. The changes are not major, but there is potential for errors
creeping in and so I recommend that everyone at least briefly look
over this document to familiarise themselves with how this looks, so
that, at minimum, people can recognise the affected code and know to
pay a little more attention.

Any feedback to myself, or Tom Clune at GSFC (Thomas.L.Clune@nasa.gov)