Regarding Drew Shindell tracers: formerly, the troposphere/stratosphere boundary for tracer chemistry purposes, etc. was LS1J(J) which was basically LS1 with 1 or 2 extra levels in the tropics.  This update switches to LTROPO(I,J), the model-calculated tropopause. This effects the following codes: TRCHEM_master.f TRCHEM_fastj.f TRCHEM_family.f TRCHEM_calc.f TRCHEM_Shindell_COM.f as well as the more general codes: TRACER_PRT.f TRACERS_DRV.f (though there should only be a change in results for Shindell tracers.)
WARNING: while we did run a multi-year test with this tropopause definition,
and things looked fine, we did not do rigorous tests (for example no tests with
chemical tracers coupled with aerosol tracers). So, please keep your eyes
peeled for anything that looks fishy near the tropopause.

Note that in the conservation quantities for Ox and NOx, the looping is now over
L=1,LS1 not L=1,LTROPO(I,J) and not L=1,LS1J(J).  So the DYNAMICS
conservation quantity signifies a flux across this nominal tropopause
(LS1/LS1-1 boundary).