cleaning up of code to remove non-standard fortran constructions.
I have turned off warnings about the number of continuation lines
being greater than 19 (so these remain).

Most changes are benign and do not effect anything, however, there
were a couple of parallel errors in the coupled code (OCNDYN.f) and
the isotope tracers (SURFACE.f). Note that the current compile options
let you know if the compiler is suspicious about a shared variable
inside a DO PARALLEL loop. If the suspicion is not justified, you can
specifically declare a variable SHARED and that will remove the warning
message.