Adjustment of tracer interaction with radiation:
I have defined a new database parameter "rad_interact_tr" to control
whether the tracers interact radiatively with the model. By default
this is set to 0 (i.e. no tracer interaction in the radiation). To
turn it on, set 'rad_interact_tr=1' in the rundeck.

Currently this is only functional for Dorothy's aerosol tracers, but
it could be extended to deal with dust, other aerosols and GHGs. We
might want to make it a bit more flexible so that different tracers
can be radiatively turned on and off from the rundeck.

I have also made some of the interactive code more generic. For
instance, the radiative forcing diagnostics. This should work fine as
more stuff gets added.

Dorothy's code is also more protected, so that TRACERS_AEROSOL_Koch
can be defined, even if her tracers are turned off (for Nadine).

Please note that all tracer ID numbers (ie. n_SO4, n_Water etc.) must
be initialised =0 so that they can act as a test of whether a tracer
is defined or not.