Usual modifications for ESMF conversion except for one hitch (below)   - modified loop bounds for J loops   - modified array bounds/allocations for JM extents.
The new awkwardness is the calls to CHECK3().  It is probably
best not to pass halo regions (which might not be initialized) to
such a check routine.  For now we've been conservative and passed
the interior (I_0:I_1,J_0:J_1), but this probably hurts performance
and is not aesthetically pleasing.   Comments/suggestions?