(1) Temporarily removing Thomas Clune's dynamic allocation of the dimensions of the "change" variable, so that the model will compile with Drew's tracers. This code is left in as comments, and will be put back in service (if needed) when the change variable coding is rewritten to be a local variable.
(2) Numerous changes that should only effect models that have SHINDELL_STRAT_CHEM
defined.  The status of the strat chem is that it is now run-able,
but still needs testing and tweaking of the chemistry. (except I still
need to update rundecks with new input files.)

(3) Otherwise, just some cleanup and maybe changing some constants to
double-precision.