[original code obtained from H. Bian (UCI) 8/2002]
This is a new code that will eventually replace TRCHEM_fastj.f, the photolysis
scheme.  For now, it is used when SHINDELL_STRAT_CHEM is defined (and therefore
should be compiled in your rundeck). If only tropospheric chemistry is used, stick
with the TRCHEM_fastj.f.  There is a check in the code in case you accidentally
try to use fastj2 with SHINDELL_STRAT_CHEM not defined.  (I am keeping both codes
in active duty, since the new one is untested and we want to be able to run the
tropospheric chemistry with the stable version in mean time.)