TRCHEM_master.f, TRCHEM_calc.f, TRCHEM_Shindell_COM.f:
Enormous amount of individual changes, but all relating to:

(1) removed the 4D change(IM,JM,LM,NTM) array and replaced it
with the 2D changeL(LM,NTM).  Originally, I was only going to
remove this from the subroutine chemstep, but it turns out it
didn't even have to be 4-dimensional in masterchem, since it
gets saved in the tr3Dsource( ) array anyway.  Thanks Gavin,
for pointing it out. However, Drew vaguely remembers that this
was made 4D in model II for the convenience of a single
diagnostic. I didn't notice that in modelE, but we should keep
an eye on our tracer diags when using this new code, just in case.

(2) making real constants double precision (X.dX format), mostly in
TRCHEM_calc.f.

After running for one month (JAN) with these two changes and then
comparing to 1 month from the model before these updates,
differences in global "change due to chemistry" conservation
quantities are less than 1% for most tracers. It was ~3% for N2O5
and PAN, ~4% for HO2NO2, and ~8% for AlkylNit.

TRCHEM_master.f: also has a fix in the strat chemistry code to
removed a dependence upon the model vertical grid.  Note that
the stratospheric chemisty code is currently giving an error.