Fix a couple of aerosol/radiation problems:
1) MDABAK=0 and MADBAK=1+FGOLDH(1:5)=0. are equivalent. Thus model
fields are now the same for tracer_aerosol_koch and non-tracer models.

2) FGOLDH(5+8) has been split into FGOLDH(5) and FTRACER(8). This is
for clarity and simplicity of parallelisation. FGOLDH(1:5) is now
global (i.e. set once in init), while FTRACER is local (and set
seperately for each I,J) to allow for radiative forcing calcs.

3) Parallel problems now gone (for TRACERS_AEROSOL_Koch at least).

4) Radiative forcing diagnostic is now done two different ways
depending on whether the final calculation uses tracers or not
(i.e. set everything to zero, and then turn each one on in turn, or,
set everything on, and turn each one off in turn). Coding could be
generalised a little better....

5) albedo fudge for drying up lakes is reduced in half to minimise odd
behaviour.