CHANGE TO DEFAULTS:
snowice=1    is now the default. Snow that is pushed below the water line
             now makes ice.

glmelt=1     The freshwater and energy associated with iceberg calving is
             now always applied around Antarctica and Greenland. It
             does not make any difference to the fixed SST run climatology,
             but it does change the q-fluxes.

Note that some landmasks may give a problem with glmelt=1. This is due
to a little bit of hard coding in LANDICE which could probably be
fixed with a little work....