Minor enhancements in radiation (no effect on balancing)
Jhour should only be used for printouts, else use nday,itime and dtsrc.
The corresponding substitutions are either carried out or suggested,
but not tested in the tracer routines.

The only other place where Jhour is currently used are the diurnal
diagnostics. The correct fixup would be adding a counter for each hour.
I will commit a quick fix that should work reasonably well, at least
for DTsrc=3600 or 3600/whole number.