In an attempt to support an automatic stack adjustment, the script "RUNID" was made a little more intelligent. Now it accepts the following optinos:    -q           - no output to prt file    -l filename  - output to the file "filename"    -i filename  - read rundeck data from "filename" (default "I")    -r           - running automatic restart (don't do "runpmE")
These were necessary so that modified runpmE could use the same
"RUNID" script to restart the model with smaller timestep.

New runpmE was not copied to /u/exec, since new runpmE requires
"RUNID" made with new setup and will not be compatible with the
runs currently running. It should be updated in /u/exec once all
current runs have finished.