first moves towards proper (i.e. resolution independent) units in the tracer jl diags. This is more complicated than I thought it would be, and so I have only made changes for the TRACERS_COSMO special jl diags. Anyone can use the functionality if required though.
To go from kg/s to kg/(m^2 mb s) change jls_jwt(k)=3 and adjust the
scaling (an extra multiple of 1d13 is generally required - look at the
conserve diags). The last line on these plots is the vertical integral
(i.e kg/(m^2 s)) and so is scaled by 1d-3 to be roughly the same order
of magnitude as the individual elements in the table. As a check, the
last line should be the same as the conservation diags.

The coding in jlmap_t is a mess. sorry.