Tracers:
Consolidation of gravitational settling at surface with turbulent
deposition. This is now consistently done in PBL, rather than TRGRAV
which now only does settling above surface.

Fixed some non-conservation and parallel problems in TRGRAV. Also
returned to use dynamic viscosity for stokes' velocity.

Note that if TRACERS_DRYDEP is not used, there will be NO dry fluxes
to the surface. If it is on, then gravitational settling is only used
if trradius(n) > 0.

Some consequent diagnostic changes:

- Conservation diag for GRAV SET now moves to a delta and is dealt
  with as part of the dry dep code. itcon_grav obsolete, replaced with
  itcon_dd(ntm,2). Separate accounting is maintained for settling
  (GRAV SET) and turbulent depostion (TURB DEP).

- trgrdep array now obsolete (flux included in trdrydep instead)

- New ij diag for grav settling for all relevant tracers (using
  taijn(tij_gsdep)). Used in calculating ij percent dry dep, but not
  included in ij turbulent dry dep flux. Total flux is now accounted
  for by adding the three different diags (tij_drydep + tij_gsdep +
  tij_prec).

Some consequent changes to dust tracers (not sure if they are correct
though).

It is conceivable that some array indicies may need to be increased
with various configurations. Please check.

Some cleaning up of code in TRACERS_DRV.