It appears that the ozone files and/or the routines interpolating those data don't give the intended results. Investigation is still going on.
Currently, monthly means are read in for selected years 1850,1890,1910,
...,1980,1990. Monthly means are interpolated for intermediate years
or extrapolated from 1980 and 1990 for years beyond 1990 linearly in
time in the stratosphere and proportional to the CH4 concentration in
the troposphere. In addition, an intra-decadal and post-1990 trend for
the stratospheric zonal means is superimposed. These routines do not
work for years prior to 1850 !

As a first step a data base parameter, O3YR_max , is added that may be
used to avoid extrapolation too far into the future. Ozone data from
year O3YR_max will be used for any later year in a cyclical fashion.
The current default is set to 1997.

If O3_yr<0 , ozone data from year -O3_yr will be used in a cyclical way.
(Using O3_yr=-1980 rather than O3_yr=1980 will make a difference only
 in the 1st half of January and the last half of December).

More changes are anticipated.